Better drag and drop (#207)
* consistent focus handling in users

up until now, users had their focus handled by themselves. It has been
changed to be more consistent with how focus should be handled across
all the components.

* Better drag and drop

* Proper menu items displayed now

CanvasStore is now responsible for showing which object actions (for
users, nodes and comments) should be visible when, depending on a
selection.