Suggestions and duplication (#254)
* Refactoring

* suggestions

* Basic walkthrough with a stepped wizard

* Improve how suggestions are rendered

* store map update where necessary

* don't store the name

* Add a node to the workspace, too

* fix model leaking from connections

* getId method

I’m tired of writing it again and again

* server-side work in progress

* ui work in progress

* Deleting a map

* Work in progress

I continue to make nodes a primary object of the workspace. It is now
possible to create nodes and establish connections, delete and edit
them as well.

* Fix deleting components

* remove sys out

* Fix improper map removal

Maps and nodes have to be removed from the workspace. Dependencies need
to be updated. This was quite expensive.

* jsPlumb cannot handle undefined label

if it is undefined, set it to the empty string

* Fix the add new node dialog

* Revert part of previous change

When referencing an existing component, it is NOT important what type
it is of.

* Referencing a node - work in progress

But unfortunately some bugs discovered:
* deleting a node from the map persists a dependency from that map, it
is necessary to remove at least one ‘visibleOn’
* reading the node duplicates nodes (possibly), error from react about
components with a duplicated keys

* clean up outgoing dependency when removing a map

* Improve jsPlumb offset calculation

* Prevent node duplication when establish a reference

* work in progress - submaps

* impact analysis passes the test

* Refactoring

* Form a submap with a review - work in progress

* Fix initial connection rendering

* prevent creation of duplicated nodes

Unregister stores before they are destructed - for some reason they
remain attached and we get multiple calls after maps is displayed twice.

* Better info about what submaps will be affected

* Fix how dialog changes

* Nodes no longer belong to a map.

It actually makes little sense to keep multiple bidirectional
references. So, from now on, a map has a *virtual* field that fetches
nodes when necessary.

* Some duplication support code

* Removed a bunch of irritating warnings.

* remove a warning

* Autofocus the dialog

* test improvements

* fix deleting a map

* Fix forming a submap

* Work in progress

* Make it possible to convert maps into submaps

* Display a lot of info about a node

Inconsistent dependencies.
Submaps.
Owners.
Etc. Etc.

* Fix q import

* Server-side controlled notification about changes. WIP.

* Better change notification framework

* Display submap suggestions, too.

* Reference existing submaps

* turn a node into submit

* migration to new history - work in progress

* Remove variants from amplest

* Reorder methods

* remove some variants server side

* Prevent dialog from reopening

Propagating event is passed back to the new map lists, which opens the
dialog immediately after it is closed.

* Fix adding a map

* small ui fixes

* Improve time handling

* New approach to the history

* remove some variants code

* Simplify schema - map adding and removal

* history and schema simplification

* prepare for component duplication

* Duplication analysis

* new duplication approach

* bugfix

events was assigned to a wrong parameter

* Fixes